refactor: split native tests out
