+++ packages/ad/__tests__/android/dom-context.android.test.js
@@ -1,4 +1,4 @@
-import shared from "../share.native";
+import shared from "../dom-context.native";
 
 describe("android", () => {
   shared();
