--- packages/ad/__tests__/ios/ad.ios.test.js
@@ -1,5 +1,0 @@
-import shared from "../share.native";
-
-describe("ios", () => {
-  shared();
-});
