+++ packages/ad/__tests__/ios/dom-context.ios.test.js
@@ -1,0 +1,5 @@
+import shared from "../dom-context.native";
+
+describe("ios", () => {
+  shared();
+});
